requirements.txt
Django
Pillow
nose
numpy
scipy
sympy
pandas
ipython
jupyter
sklearn
matplotlib
psycopg2-binary
djangorestframework
